requirements.txt
faiss-cpu
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
openai
pandas
plotly
pypdf
python-dotenv
pysqlite3-binary
streamlit
